Added Rate entity and one to many relation with ClassEmployee, created CRUD operations for Rate

+++ javaHibernate/JavaJdbcHibernate/src/main/java/org/ClassEmployeeController.java
@@ -3,6 +3,7 @@ package org;
 import javax.persistence.*;
 import java.util.HashSet;
 import java.util.List;
+import java.util.Optional;
 import java.util.Set;
 
 public class ClassEmployeeController {
@@ -26,6 +27,7 @@ public class ClassEmployeeController {
             ce = tq.getSingleResult();
             System.out.println(ce.getClassName() + ", " + ce.getMaxNum());
         } catch(NoResultException e) {
+            System.out.println("Class with that id not found.\n");
             e.printStackTrace();
         } finally {
             em.close();
@@ -87,13 +89,18 @@ public class ClassEmployeeController {
 
             ce = em.find(ClassEmployee.class, id);
 
-            // Create a copy of the employee collection to avoid ConcurrentModificationException
-            Set<Employee> employeesCopy = new HashSet<>(ce.getEmployees());
-            for (Employee employee : employeesCopy) {
-                employee.removeClassEmployee(ce);
-            }
+            if (ce != null){
+                Set<Employee> employeesCopy = new HashSet<>(ce.getEmployees());
+                for (Employee employee : employeesCopy) {
+                    employee.removeClassEmployee(ce);
+                }
 
-            em.remove(ce);
+                em.remove(ce);
+            }
+            else {
+                System.out.println("Class with that id not found.\n");
+            }
+            // Create a copy of the employee collection to avoid ConcurrentModificationException
 
             et.commit();
         }
@@ -104,6 +111,38 @@ public class ClassEmployeeController {
             e.printStackTrace();
         }
         finally {
+            em.close();
+        }
+    }
+
+    public static void updateClassEmployee(int id, Optional<String> newClassName, Optional<Integer> newMaxNum) {
+        EntityManager em = ENTITY_MANAGER_FACTORY.createEntityManager();
+        EntityTransaction et = null;
+
+        try {
+            et = em.getTransaction();
+            et.begin();
+
+            ClassEmployee ce = em.find(ClassEmployee.class, id);
+
+            if (ce != null) {
+                // Update the classEmployee details if the corresponding optional values are present
+                newClassName.ifPresent(ce::setClassName);
+                newMaxNum.ifPresent(ce::setMaxNum);
+
+                em.persist(ce);
+                et.commit();
+
+                System.out.println("ClassEmployee updated successfully.");
+            } else {
+                System.out.println("Employee with ID " + id + " not found.");
+            }
+        } catch (Exception e) {
+            if (et != null) {
+                et.rollback();
+            }
+            e.printStackTrace();
+        } finally {
             em.close();
         }
     }
